siprems-backend/requirements.txt
flask
pandas
prophet
flask-cors
psycopg2-binary
python-dotenv
google-generativeai
joblib
scikit-learn
pyjwt
werkzeug
flask-talisman
flask-limiter
marshmallow
